Rename references to challenge server to andromeda
