+++ src/backend/urls.py
@@ -25,7 +25,7 @@ urlpatterns = [
     path('announcements/', include('announcements.urls')),
     path('auth/', include('authentication.urls')),
     path('challenges/', include('challenge.urls')),
-    path('challengeserver/', include('challengeserver.urls')),
+    path('challengeserver/', include('andromeda.urls')),
     path('config/', include('config.urls')),
     path('hints/', include('hint.urls')),
     path('leaderboard/', include('leaderboard.urls')),
